COMMERCE-42 Regen services

+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPAttachmentFileEntryLocalService.java
@@ -88,7 +88,7 @@ public interface CPAttachmentFileEntryLocalService extends BaseLocalService,
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		Map<Locale, java.lang.String> titleMap, java.lang.String json,
-		int priority, int type, ServiceContext serviceContext)
+		double priority, int type, ServiceContext serviceContext)
 		throws PortalException;
 
 	/**
@@ -179,7 +179,7 @@ public interface CPAttachmentFileEntryLocalService extends BaseLocalService,
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		Map<Locale, java.lang.String> titleMap, java.lang.String json,
-		int priority, int type, ServiceContext serviceContext)
+		double priority, int type, ServiceContext serviceContext)
 		throws PortalException;
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
