+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPAttachmentFileEntryLocalServiceWrapper.java
@@ -54,7 +54,7 @@ public class CPAttachmentFileEntryLocalServiceWrapper
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		java.lang.String json, int priority, int type,
+		java.lang.String json, double priority, int type,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpAttachmentFileEntryLocalService.addCPAttachmentFileEntry(classNameId,
@@ -175,7 +175,7 @@ public class CPAttachmentFileEntryLocalServiceWrapper
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		java.lang.String json, int priority, int type,
+		java.lang.String json, double priority, int type,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpAttachmentFileEntryLocalService.updateCPAttachmentFileEntry(cpAttachmentFileEntryId,
