+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPAttachmentFileEntryServiceWrapper.java
@@ -42,7 +42,7 @@ public class CPAttachmentFileEntryServiceWrapper
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		java.lang.String json, int priority, int type,
+		java.lang.String json, double priority, int type,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpAttachmentFileEntryService.addCPAttachmentFileEntry(classNameId,
@@ -89,7 +89,7 @@ public class CPAttachmentFileEntryServiceWrapper
 		int expirationDateYear, int expirationDateHour,
 		int expirationDateMinute, boolean neverExpire,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		java.lang.String json, int priority, int type,
+		java.lang.String json, double priority, int type,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpAttachmentFileEntryService.updateCPAttachmentFileEntry(cpAttachmentFileEntryId,
