+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPDefinitionOptionRelService.java
@@ -74,9 +74,10 @@ public interface CPDefinitionOptionRelService extends BaseService {
 		long cpOptionId, java.lang.String name,
 		Map<Locale, java.lang.String> titleMap,
 		Map<Locale, java.lang.String> descriptionMap,
-		java.lang.String ddmFormFieldTypeName, int priority, boolean facetable,
-		boolean required, boolean skuContributor, boolean importOptionValue,
-		ServiceContext serviceContext) throws PortalException;
+		java.lang.String ddmFormFieldTypeName, double priority,
+		boolean facetable, boolean required, boolean skuContributor,
+		boolean importOptionValue, ServiceContext serviceContext)
+		throws PortalException;
 
 	public CPDefinitionOptionRel deleteCPDefinitionOptionRel(
 		CPDefinitionOptionRel cpDefinitionOptionRel) throws PortalException;
@@ -106,9 +107,9 @@ public interface CPDefinitionOptionRelService extends BaseService {
 		long cpDefinitionOptionRelId, long cpOptionId, java.lang.String name,
 		Map<Locale, java.lang.String> titleMap,
 		Map<Locale, java.lang.String> descriptionMap,
-		java.lang.String ddmFormFieldTypeName, int priority, boolean facetable,
-		boolean required, boolean skuContributor, ServiceContext serviceContext)
-		throws PortalException;
+		java.lang.String ddmFormFieldTypeName, double priority,
+		boolean facetable, boolean required, boolean skuContributor,
+		ServiceContext serviceContext) throws PortalException;
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public BaseModelSearchResult<CPDefinitionOptionRel> searchCPDefinitionOptionRels(
