+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPDefinitionOptionValueRelLocalServiceWrapper.java
@@ -60,7 +60,7 @@ public class CPDefinitionOptionValueRelLocalServiceWrapper
 	public com.liferay.commerce.product.model.CPDefinitionOptionValueRel addCPDefinitionOptionValueRel(
 		long cpDefinitionOptionRelId, java.lang.String name,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		int priority,
+		double priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpDefinitionOptionValueRelLocalService.addCPDefinitionOptionValueRel(cpDefinitionOptionRelId,
@@ -173,7 +173,7 @@ public class CPDefinitionOptionValueRelLocalServiceWrapper
 	public com.liferay.commerce.product.model.CPDefinitionOptionValueRel updateCPDefinitionOptionValueRel(
 		long cpDefinitionOptionValueRelId, java.lang.String name,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		int priority,
+		double priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpDefinitionOptionValueRelLocalService.updateCPDefinitionOptionValueRel(cpDefinitionOptionValueRelId,
