+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPDefinitionOptionValueRelService.java
@@ -64,7 +64,7 @@ public interface CPDefinitionOptionValueRelService extends BaseService {
 	 */
 	public CPDefinitionOptionValueRel addCPDefinitionOptionValueRel(
 		long cpDefinitionOptionRelId, java.lang.String name,
-		Map<Locale, java.lang.String> titleMap, int priority,
+		Map<Locale, java.lang.String> titleMap, double priority,
 		ServiceContext serviceContext) throws PortalException;
 
 	public CPDefinitionOptionValueRel deleteCPDefinitionOptionValueRel(
@@ -84,7 +84,7 @@ public interface CPDefinitionOptionValueRelService extends BaseService {
 
 	public CPDefinitionOptionValueRel updateCPDefinitionOptionValueRel(
 		long cpDefinitionOptionValueRelId, java.lang.String name,
-		Map<Locale, java.lang.String> titleMap, int priority,
+		Map<Locale, java.lang.String> titleMap, double priority,
 		ServiceContext serviceContext) throws PortalException;
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
