+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPDefinitionOptionValueRelServiceWrapper.java
@@ -38,7 +38,7 @@ public class CPDefinitionOptionValueRelServiceWrapper
 	public com.liferay.commerce.product.model.CPDefinitionOptionValueRel addCPDefinitionOptionValueRel(
 		long cpDefinitionOptionRelId, java.lang.String name,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		int priority,
+		double priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpDefinitionOptionValueRelService.addCPDefinitionOptionValueRel(cpDefinitionOptionRelId,
@@ -77,7 +77,7 @@ public class CPDefinitionOptionValueRelServiceWrapper
 	public com.liferay.commerce.product.model.CPDefinitionOptionValueRel updateCPDefinitionOptionValueRel(
 		long cpDefinitionOptionValueRelId, java.lang.String name,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		int priority,
+		double priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _cpDefinitionOptionValueRelService.updateCPDefinitionOptionValueRel(cpDefinitionOptionValueRelId,
