+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CPOptionCategoryService.java
@@ -60,7 +60,7 @@ public interface CPOptionCategoryService extends BaseService {
 	 */
 	public CPOptionCategory addCPOptionCategory(java.lang.String name,
 		Map<Locale, java.lang.String> titleMap,
-		Map<Locale, java.lang.String> descriptionMap, int priority,
+		Map<Locale, java.lang.String> descriptionMap, double priority,
 		ServiceContext serviceContext) throws PortalException;
 
 	public CPOptionCategory deleteCPOptionCategory(
@@ -79,7 +79,7 @@ public interface CPOptionCategoryService extends BaseService {
 
 	public CPOptionCategory updateCPOptionCategory(long cpOptionCategoryId,
 		java.lang.String name, Map<Locale, java.lang.String> titleMap,
-		Map<Locale, java.lang.String> descriptionMap, int priority,
+		Map<Locale, java.lang.String> descriptionMap, double priority,
 		ServiceContext serviceContext) throws PortalException;
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
